Move nav button and fix navigation orbit
